# openapi.json
{"openapi":"3.1.0","info":{"title":"GlobalRoamer Enterprise AI Observability Platform","version":"0.1.0"},"paths":{"/live":{"get":{"tags":["health"],"summary":"Liveness check","description":"Confirm that the API process is running.\n\nThis endpoint must not depend on PostgreSQL or any external service.","operationId":"liveness_live_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Liveness Live Get"}}}}}}},"/ready":{"get":{"tags":["health"],"summary":"Readiness check","description":"Confirm that the application is ready to receive traffic.\n\nThe application is considered ready only when required\ninfrastructure dependencies are available.","operationId":"readiness_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"503":{"description":"Application is not ready"}}}},"/health":{"get":{"tags":["health"],"summary":"Application health check","description":"Return the aggregated health status of the application.\n\nAdditional components such as the vector database, workers,\nfilesystem and external AI providers can be added later.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"503":{"description":"Application is unhealthy"}}}},"/api/v1/traces":{"post":{"tags":["traces"],"summary":"Create a trace","description":"Create a new tenant-scoped trace.","operationId":"create_trace_api_v1_traces_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTraceRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TraceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["traces"],"summary":"List traces","description":"Retrieve traces using optional tenant filtering and pagination.","operationId":"list_traces_api_v1_traces_get","parameters":[{"name":"tenant_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string","minLength":1,"maxLength":100},{"type":"null"}],"description":"Filter traces by tenant identifier.","title":"Tenant Id"},"description":"Filter traces by tenant identifier."},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"description":"Maximum number of traces to return.","default":100,"title":"Limit"},"description":"Maximum number of traces to return."},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"description":"Number of traces to skip.","default":0,"title":"Offset"},"description":"Number of traces to skip."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TraceListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/traces/{trace_id}":{"get":{"tags":["traces"],"summary":"Get a trace","description":"Retrieve one trace by its internal UUID.","operationId":"get_trace_api_v1_traces__trace_id__get","parameters":[{"name":"trace_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Trace Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TraceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/traces/process":{"post":{"tags":["Trace Processing"],"summary":"Process and persist a trace","description":"Loads a trace artifact, parses and normalizes its contents, extracts evidence, signals, and values, applies configured mappings, and persists the parsed trace in PostgreSQL.","operationId":"process_trace_api_v1_traces_process_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProcessTraceRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProcessTraceResponse"}}}},"400":{"description":"The trace-processing request or source artifact is invalid."},"404":{"description":"The requested source artifact does not exist."},"500":{"description":"The trace-processing workflow failed."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/traces/submit":{"post":{"tags":["Trace Submission"],"summary":"Submit Trace","description":"Accept a server-side trace file for asynchronous processing.","operationId":"submit_trace_api_v1_traces_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitTraceRequest"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitTraceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CreateTraceRequest":{"properties":{"tenant_id":{"type":"string","maxLength":100,"minLength":1,"title":"Tenant Id"},"trace_id":{"type":"string","maxLength":128,"minLength":1,"title":"Trace Id"},"testcase_id":{"type":"string","maxLength":128,"minLength":1,"title":"Testcase Id"}},"type":"object","required":["tenant_id","trace_id","testcase_id"],"title":"CreateTraceRequest","description":"Request payload for creating a trace."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ProcessTraceRequest":{"properties":{"source_path":{"type":"string","maxLength":4096,"minLength":1,"title":"Source Path","description":"Path to a trace file available inside the API runtime."},"tenant_id":{"type":"string","maxLength":100,"minLength":1,"title":"Tenant Id","description":"Tenant that owns the trace."},"trace_id":{"type":"string","maxLength":128,"minLength":1,"title":"Trace Id","description":"External identifier of the trace."},"testcase_id":{"anyOf":[{"type":"string","maxLength":128,"minLength":1},{"type":"null"}],"title":"Testcase Id","description":"Optional testcase identifier associated with the trace."}},"additionalProperties":false,"type":"object","required":["source_path","tenant_id","trace_id"],"title":"ProcessTraceRequest","description":"Request payload for processing and persisting one trace file.","example":{"source_path":"etc/sample_trace.csv","tenant_id":"smoke-test","testcase_id":"process-trace-001","trace_id":"process-trace-001"}},"ProcessTraceResponse":{"properties":{"parsed_trace_id":{"type":"string","format":"uuid","title":"Parsed Trace Id","description":"Identifier of the persisted parsed-trace record."},"tenant_id":{"type":"string","title":"Tenant Id","description":"Tenant that owns the processed trace."},"trace_id":{"type":"string","title":"Trace Id","description":"External trace identifier."},"testcase_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Testcase Id","description":"Optional testcase identifier associated with the trace."},"row_count":{"type":"integer","minimum":0.0,"title":"Row Count","description":"Number of rows loaded from the trace."},"evidence_count":{"type":"integer","minimum":0.0,"title":"Evidence Count","description":"Number of evidence records extracted."},"signal_count":{"type":"integer","minimum":0.0,"title":"Signal Count","description":"Number of signals extracted."},"extracted_value_count":{"type":"integer","minimum":0.0,"title":"Extracted Value Count","description":"Number of values extracted from the trace."},"mapped_value_count":{"type":"integer","minimum":0.0,"title":"Mapped Value Count","description":"Number of values mapped successfully."},"warning_count":{"type":"integer","minimum":0.0,"title":"Warning Count","description":"Number of processing warnings."},"error_count":{"type":"integer","minimum":0.0,"title":"Error Count","description":"Number of processing errors."},"is_valid":{"type":"boolean","title":"Is Valid","description":"Whether the resulting parsed trace passed validation."},"is_complete":{"type":"boolean","title":"Is Complete","description":"Whether processing completed with all required data."}},"additionalProperties":false,"type":"object","required":["parsed_trace_id","tenant_id","trace_id","row_count","evidence_count","signal_count","extracted_value_count","mapped_value_count","warning_count","error_count","is_valid","is_complete"],"title":"ProcessTraceResponse","description":"Response returned after successful trace processing and persistence.","example":{"error_count":0,"evidence_count":3,"extracted_value_count":9,"is_complete":true,"is_valid":true,"mapped_value_count":6,"parsed_trace_id":"9d890946-cb38-44f1-818c-c6bb16838833","row_count":3,"signal_count":9,"tenant_id":"smoke-test","testcase_id":"process-trace-001","trace_id":"process-trace-001","warning_count":0}},"SubmitTraceRequest":{"properties":{"source_path":{"type":"string","maxLength":4096,"minLength":1,"title":"Source Path"},"tenant_id":{"type":"string","maxLength":100,"minLength":1,"title":"Tenant Id"},"trace_id":{"type":"string","maxLength":128,"minLength":1,"title":"Trace Id"},"testcase_id":{"anyOf":[{"type":"string","maxLength":128,"minLength":1},{"type":"null"}],"title":"Testcase Id"}},"additionalProperties":false,"type":"object","required":["source_path","tenant_id","trace_id"],"title":"SubmitTraceRequest","description":"Request for submitting an existing server-side trace file."},"SubmitTraceResponse":{"properties":{"artifact_id":{"type":"string","format":"uuid","title":"Artifact Id"},"submission_event_id":{"type":"string","format":"uuid","title":"Submission Event Id"},"outbox_message_id":{"type":"string","format":"uuid","title":"Outbox Message Id"},"tenant_id":{"type":"string","title":"Tenant Id"},"trace_id":{"type":"string","title":"Trace Id"},"testcase_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Testcase Id"},"correlation_id":{"type":"string","title":"Correlation Id"},"storage_key":{"type":"string","title":"Storage Key"},"status":{"type":"string","title":"Status"}},"additionalProperties":false,"type":"object","required":["artifact_id","submission_event_id","outbox_message_id","tenant_id","trace_id","testcase_id","correlation_id","storage_key","status"],"title":"SubmitTraceResponse","description":"Accepted trace-submission response."},"TraceListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/TraceResponse"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["items","total"],"title":"TraceListResponse","description":"Paginated trace collection."},"TraceResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"tenant_id":{"type":"string","title":"Tenant Id"},"trace_id":{"type":"string","title":"Trace Id"},"testcase_id":{"type":"string","title":"Testcase Id"},"status":{"type":"string","title":"Status"},"current_stage":{"type":"string","title":"Current Stage"},"version":{"type":"integer","title":"Version"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","tenant_id","trace_id","testcase_id","status","current_stage","version","created_at","updated_at"],"title":"TraceResponse","description":"API representation of a trace."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}